Fix(getAvtiveMonthsRanges): Fixed ranges to account for overflow at end of year
Active months that continue over the end of year are now accounted for in the range. I.e. for a critter available October to February, we previously had [[0,2],[10,11]] but now have [[10,2]].

This fixed a bug in when filtering critters by 'last chance' and 'new this month' - previously, the critter from above would have been marked as 'last chance' in december even though it would be available in January.