Fix(Redux): Fixed state bugs when closing and re-opening app
Bug: If state was changed (i.e. a fish added to 'donated fish') and the app closed. When app reopened, 'fish' would still be in 'donatedFish' list, however filtering/sorting/highlighting the donated button were all functioning incorrectly. Donated button would indicate fish was not donated even though it was. Filters/sort would not indicate fish was in donated list.

Fix: This was due to using Array.includes(). This passes by reference not value, and hence when checking if a fish was in donatedFish, includes() would return nothing as it was not the same instance of a fish, although the fish were functionlly the same (same name etc). Changed to compare by values, i.e. fish.name === donated.name.